backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
google-generativeai
google-genai
pydantic
Pillow
httpx[socks]